+++ src/Models/AdaMultiLasso.cpp
@@ -37,7 +37,7 @@ AdaMultiLasso::AdaMultiLasso(const unordered_map<string, string>& opts) {
     initTrainingFlag = false;
 }
 
-void AdaMultiLasso::setAttributeMatrix(const string& str, MatrixXd* Z) {
+void AdaMultiLasso::setAttributeMatrix(const string& str, MatrixXf* Z) {
     if (str == "snpsFeatures1" || str == "snpsFeature1") {
         setSnpsFeatures1(*Z);
     } else if (str == "snpsFeatures2" || str == "snpsFeature2") {
@@ -58,24 +58,24 @@ void AdaMultiLasso::setLambda2(double d) {
     lambda2 = d;
 }
 
-void AdaMultiLasso::setSnpsFeatures1(MatrixXd xd) {
+void AdaMultiLasso::setSnpsFeatures1(MatrixXf xd) {
     snpsFeatures1 = xd;
 }
 
-void AdaMultiLasso::setSnpsFeatures2(MatrixXd xd) {
+void AdaMultiLasso::setSnpsFeatures2(MatrixXf xd) {
     snpsFeatures2 = xd;
 }
 
-void AdaMultiLasso::setSnpsFeatures(MatrixXd xd) {
+void AdaMultiLasso::setSnpsFeatures(MatrixXf xd) {
     setSnpsFeatures1(xd);
     setSnpsFeatures2(xd);
 }
 
-MatrixXd AdaMultiLasso::getSnpsFeatures1() {
+MatrixXf AdaMultiLasso::getSnpsFeatures1() {
     return snpsFeatures1;
 }
 
-MatrixXd AdaMultiLasso::getSnpsFeatures2() {
+MatrixXf AdaMultiLasso::getSnpsFeatures2() {
     return snpsFeatures2;
 }
 
@@ -168,10 +168,10 @@ VectorXd AdaMultiLasso::getRho() {
     return rho;
 }
 
-MatrixXd AdaMultiLasso::getBeta() {
+MatrixXf AdaMultiLasso::getBeta() {
     long r = beta.rows()/taskNum;
     long c = taskNum;
-    MatrixXd tmp = MatrixXd::Zero(r, c);
+    MatrixXf tmp = MatrixXf::Zero(r, c);
     for (long i=0;i<r;i++){
         for (long j=0;j<c;j++){
             tmp(i,j) = beta(i*taskNum+j, 0);
@@ -180,19 +180,19 @@ MatrixXd AdaMultiLasso::getBeta() {
     return tmp;
 }
 
-MatrixXd AdaMultiLasso::getFormattedBeta() {
+MatrixXf AdaMultiLasso::getFormattedBeta() {
     return beta;
 }
 
-void AdaMultiLasso::setX(MatrixXd xd) {
+void AdaMultiLasso::setX(MatrixXf xd) {
     X = xd;
 }
 
-void AdaMultiLasso::setY(MatrixXd yd) {
+void AdaMultiLasso::setY(MatrixXf yd) {
     y = yd;
 }
 
-void AdaMultiLasso::setXY(MatrixXd m, MatrixXd n) {
+void AdaMultiLasso::setXY(MatrixXf m, MatrixXf n) {
     X = m;
     y = n;
 }
@@ -200,7 +200,7 @@ void AdaMultiLasso::setXY(MatrixXd m, MatrixXd n) {
 void AdaMultiLasso::initBeta() {
     long c = X.cols();
     long d = y.cols();
-    beta = MatrixXd::Random(c, d);
+    beta = MatrixXf::Random(c, d);
 }
 
 double AdaMultiLasso::cost() {
@@ -209,7 +209,7 @@ double AdaMultiLasso::cost() {
 }
 
 double AdaMultiLasso::penalty_cost() {
-    MatrixXd rb = getBeta();
+    MatrixXf rb = getBeta();
     long c = rb.rows();
     double result = 0;
     VectorXd theta = getTheta()*lambda1;
@@ -248,12 +248,12 @@ void AdaMultiLasso::initTraining() {
         long n = X.rows();
         long c = X.cols();
         taskNum = y.cols();
-        MatrixXd tmpX = MatrixXd::Zero(n*taskNum, c*taskNum);
-        MatrixXd tmpY = MatrixXd::Zero(n*taskNum, 1);
+        MatrixXf tmpX = MatrixXf::Zero(n*taskNum, c*taskNum);
+        MatrixXf tmpY = MatrixXf::Zero(n*taskNum, 1);
         VectorXd tmpT = VectorXd::Zero(c*taskNum);
         VectorXd tmpR = VectorXd::Zero(c*taskNum);
-        MatrixXd tmpS1 = MatrixXd::Zero(c*taskNum, snpsFeatures1.cols());
-        MatrixXd tmpS2 = MatrixXd::Zero(c*taskNum, snpsFeatures2.cols());
+        MatrixXf tmpS1 = MatrixXf::Zero(c*taskNum, snpsFeatures1.cols());
+        MatrixXf tmpS2 = MatrixXf::Zero(c*taskNum, snpsFeatures2.cols());
         for (long j=0;j<taskNum;j++){
             for (long k=0;k<c;k++){
                 tmpT(k*taskNum+j) = theta(k);
@@ -284,7 +284,7 @@ void AdaMultiLasso::initTraining() {
 
 void AdaMultiLasso::initC() {
     long c = X.cols();
-    C = MatrixXd::Zero(c * taskNum, c);
+    C = MatrixXf::Zero(c * taskNum, c);
     for (long i = 0; i < c; i++) {
         for (long j = 0; j < taskNum; j++) {
             C(i*j+j, i) = rho(i)*lambda2;
@@ -292,19 +292,19 @@ void AdaMultiLasso::initC() {
     }
 }
 
-MatrixXd AdaMultiLasso::proximal_derivative() {
+MatrixXf AdaMultiLasso::proximal_derivative() {
     long r = beta.rows();
     long c = beta.cols();
-    MatrixXd A = MatrixXd::Zero(r*taskNum, c);
-    MatrixXd tmp = C*beta;
+    MatrixXf A = MatrixXf::Zero(r*taskNum, c);
+    MatrixXf tmp = C*beta;
     for (long i=0;i<r*taskNum;i++){
         A.row(i) = Math::getInstance().L2Thresholding(tmp.row(i)/mu);
     }
     return X.transpose()*(X*beta-y) + C.transpose()*A;
 }
 
-MatrixXd AdaMultiLasso::proximal_operator(MatrixXd in, float lr) {
-    MatrixXd sign = ((in.array()>0).matrix()).cast<double>();
+MatrixXf AdaMultiLasso::proximal_operator(MatrixXf in, float lr) {
+    MatrixXf sign = ((in.array()>0).matrix()).cast<double>();
     sign += -1.0*((in.array()<0).matrix()).cast<double>();
     in = ((in.array().abs()-lr*lambda1*theta.array()).max(0)).matrix();
     return (in.array()*sign.array()).matrix();
