+++ src/Models/SparseLMM.h
@@ -11,8 +11,8 @@
 class SparseLMM : public LinearMixedModel{
 private:
     double l1Reg;
-    MatrixXd rX;
-    MatrixXd rY;
+    MatrixXf rX;
+    MatrixXf rY;
 
     static constexpr double default_l1Reg = 0.0;
 public:
@@ -20,8 +20,8 @@ public:
     SparseLMM(const unordered_map<string, string>& options);
 
     void rotateXY(double);
-    MatrixXd getRotatedX();
-    MatrixXd getRoattedY();
+    MatrixXf getRotatedX();
+    MatrixXf getRoattedY();
 
     void setL1reg(double);
     double getL1reg();
